backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.1
Pillow==10.1.0
reportlab==4.0.7
python-decouple==3.8
openpyxl==3.1.2
pandas==2.1.4
PyPDF2==3.0.1
